[TSK-276] refactor: reorganize routes by moving Gestao components to corporativo and cleaning up unused files
