--- app/routes/Gestao/__init__.py
@@ -1,4 +1,0 @@
-from .. import cargos
-from . import departamentos, empresas, funcionarios
-
-__all__ = [cargos, empresas, funcionarios, departamentos]
