+++ app/routes/corporativo/empresas.py
@@ -1,31 +1,26 @@
-from flask import abort, render_template, request
+from flask import abort, render_template
 from flask_login import login_required
 
-from app import app
-from app.decorators import read_perm, set_endpoint
+from . import corp
 from app.forms.create import CadastroEmpresa
-from app.forms.globals import IMPORTEPIForm
 from app.models.Funcionários import Empresa
 
 
-@app.route("/Empresas", methods=["GET"])
+@corp.route("/Empresas", methods=["GET"])
 @login_required
-@set_endpoint
-@read_perm
 def Empresas():
     try:
-        importForm = IMPORTEPIForm()
+
         form = CadastroEmpresa()
         database = Empresa.query.all()
-        DataTables = f"js/DataTables/gestao/{request.endpoint.capitalize()}Table.js"
-        page = f"pages/Gestao/{request.endpoint.lower()}.html"
+        DataTables = ""
+        page = ""
         return render_template(
             "index.html",
             page=page,
             form=form,
             DataTables=DataTables,
             database=database,
-            importForm=importForm,
         )
 
     except Exception as e:
