+++ app/routes/corporativo/templates/cargos.html
@@ -75,7 +75,3 @@
   </div>
 
 </div>
-<div id="results">
-
-</div>
-{% include 'pages/forms/cargos/create.html' %}
