+++ app/routes/serving/__init__.py
@@ -1,6 +1,8 @@
 from pathlib import Path
-from flask import Blueprint, send_from_directory
+
+from flask import Blueprint
 from flask import current_app as app
+from flask import send_from_directory
 
 serve = Blueprint("serve", __name__)
